fix 64bit warning and add switch breaks to java demo

+++ cpp/libs/src/asiodnp3/IOHandler.cpp
@@ -222,7 +222,7 @@ void IOHandler::BeginRead()
 		}
 		else
 		{
-			self->parser.OnRead(num, *self);
+			self->parser.OnRead(static_cast<uint32_t>(num), *self);
 			self->BeginRead();
 		}
 	};
